69: Update SDL2 to SDL3 (#70)
* 69: Width and height in PongScene and GeometryTestScene also changed to float for consistency

+++ gp/sdl/internal/sdl_context.hpp
@@ -25,7 +25,7 @@ public:
    *
    * @return The number of milliseconds since the SDL context was created.
    */
-  std::uint32_t timestamp() const;
+  std::uint64_t timestamp() const;
   int exec() const;
   void request_close();
