+++ wolf/wolf_sprite_voxelizer/voxel_scene.cpp
@@ -88,7 +88,7 @@ void VoxelScene::resize(const int width, const int height) {
   projection_ = glm::perspective(glm::radians(60.0f), static_cast<float>(width) / height, 0.1f, 1000.0f);
 }
 
-void VoxelScene::animate(const std::uint32_t timestamp) {
+void VoxelScene::animate(const std::uint64_t timestamp) {
   const auto time_elapsed_ms = timestamp - last_timestamp_ms_;
 
   if (animate_) {
